+++ src/aui-button/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1806](https://issues.liferay.com/browse/AUI-1806) Add button to cancel editing/creating a field
 * [AUI-1793](https://issues.liferay.com/browse/AUI-1793) Fix date field bug with the boolean data editors
 * [AUI-1724](https://issues.liferay.com/browse/AUI-1724) Improve aui-boolean-data-editor's appearance
 * [AUI-1543](https://issues.liferay.com/browse/AUI-1543) Make aui-button accessible